+++ packages/theme/eslint.config.mjs
@@ -1,2 +1,11 @@
 import { config } from "@aristobyte-ui/eslint-config/react";
-export default config;
+
+export default [
+  ...config,
+  {
+    name: "custom-overrides",
+    rules: {
+      "@typescript-eslint/no-explicit-any": "off",
+    },
+  },
+];