+++ packages/theme/plugins/spinner.ts
@@ -1,16 +1,4 @@
-// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-// @ts-ignore
-import plugin from "tailwindcss/plugin";
-import type { PluginType } from "./types";
-
-export const spinner: PluginType = plugin(function ({
-  // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-  // @ts-ignore
-  addComponents,
-  // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-  // @ts-ignore
-  addBase,
-}) {
+export function spinner({ addComponents, addBase }: any) {
   addBase({
     "@keyframes rotation": {
       from: { transform: "rotate(0deg)" },
@@ -258,4 +246,4 @@ export const spinner: PluginType = plugin(function ({
     ...typeMap,
     ...variantsMap,
   });
-});
+}