+++ packages/card/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @aristobyte-ui/anchor
 
+## 1.0.99
+
+### Patch Changes
+
+- fix: remove spinner from the theme
+
+- Updated dependencies []:
+  - @aristobyte-ui/label@1.0.99
+  - @aristobyte-ui/utils@1.0.99
+
 ## 1.0.98
 
 ### Patch Changes