+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @aristobyte-ui/cli
 
+## 1.0.99
+
+### Patch Changes
+
+- fix: remove spinner from the theme
+
+- Updated dependencies []:
+  - @aristobyte-ui/eslint-config@1.0.99
+  - @aristobyte-ui/typescript-config@1.0.99
+
 ## 1.0.98
 
 ### Patch Changes