+++ packages/dropdown/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @aristobyte-ui/dropdown
 
+## 1.0.99
+
+### Patch Changes
+
+- fix: remove spinner from the theme
+
+- Updated dependencies []:
+  - @aristobyte-ui/button@1.0.99
+  - @aristobyte-ui/utils@1.0.99
+
 ## 1.0.98
 
 ### Patch Changes