+++ packages/eslint-config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @aristobyte-ui/eslint-config
 
+## 1.0.99
+
+### Patch Changes
+
+- fix: remove spinner from the theme
+
 ## 1.0.98
 
 ## 1.0.97