+++ packages/message-box/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @aristobyte-ui/message-box
 
+## 1.0.99
+
+### Patch Changes
+
+- fix: remove spinner from the theme
+
+- Updated dependencies []:
+  - @aristobyte-ui/utils@1.0.99
+
 ## 1.0.98
 
 ### Patch Changes